Regroup module, function, and decorator tests (wala/ML#635 Phase 3).
`TestModules` takes the 92 module- and import-binding tests,
`TestFunctions` the 15 function-dispatch/redefinition tests, and the 12
`testDecorator*` methods fold into the existing `TestDecoratedMethods`;
the monolith keeps 278, preserving the module total.
